refactor(web): extract useTimerCalculations hook
DisplayScreen interleaved ~100 lines of timer math (gross/net/saldo,
legal-pause tier placement, segment building, marker angles,
contextual message) with JSX. Move the whole computation into a
dedicated hook and let DisplayScreen focus on composition.

DisplayScreen: 175 → 64 LoC. BreakHoverItem now lives in the hook
(the producer) and BreakTooltip / TimerRing import the type from
there.
